Moved all email parsing to Vaulty lib
