+++ vaulty-mail/filter/Cargo.toml
@@ -9,7 +9,6 @@ vaulty = { path = "../lib" }
 env_logger = "0.7.1"
 log = "0.4.8"
 structopt = "0.3.9"
-mailparse = "0.10.2"
 reqwest = { version = "0.10.1", features = ["blocking"] }
 serde = { version = "1.0.104", features = ["derive"] }
 serde_json = "1.0.44"
