+++ vaulty-mail/lib/Cargo.toml
@@ -12,6 +12,7 @@ url = "2.1.0"
 log = "0.4.8"
 chrono = "0.4.10"
 bytes = "0.5.3"
+mailparse = "0.10.2"
 
 [dev-dependencies]
 tokio = { version = "0.2.6", features = ["full"] }
